requirements.txt
gitpython == '2.1.8'